// web/src/generated/api/models/ParticipantResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cuentas Claras API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Current participant identity and archive state.
 * @export
 * @interface ParticipantResponse
 */
export interface ParticipantResponse {
    /**
     * 
     * @type {boolean}
     * @memberof ParticipantResponse
     */
    archived: boolean;
    /**
     * 
     * @type {Date}
     * @memberof ParticipantResponse
     */
    createdAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof ParticipantResponse
     */
    groupId: string;
    /**
     * 
     * @type {string}
     * @memberof ParticipantResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ParticipantResponse
     */
    name: string;
}

/**
 * Check if a given object implements the ParticipantResponse interface.
 */
export function instanceOfParticipantResponse(value: object): value is ParticipantResponse {
    if (!('archived' in value) || value['archived'] === undefined) return false;
    if (!('groupId' in value) || value['groupId'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function ParticipantResponseFromJSON(json: any): ParticipantResponse {
    return ParticipantResponseFromJSONTyped(json, false);
}

export function ParticipantResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ParticipantResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'archived': json['archived'],
        'createdAt': json['created_at'] == null ? undefined : (new Date(json['created_at'])),
        'groupId': json['group_id'],
        'id': json['id'],
        'name': json['name'],
    };
}

export function ParticipantResponseToJSON(json: any): ParticipantResponse {
    return ParticipantResponseToJSONTyped(json, false);
}

export function ParticipantResponseToJSONTyped(value?: ParticipantResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'archived': value['archived'],
        'created_at': value['createdAt'] === null ? null : ((value['createdAt'] as any)?.toISOString()),
        'group_id': value['groupId'],
        'id': value['id'],
        'name': value['name'],
    };
}